attempt to implement rest of oled funcs
...adding my thoughts on how we could do it where it is not obvious

+++ ByggernKybInf/drivers/oled.h
@@ -37,7 +37,7 @@ void OLED_print_char(char c);
 
 void OLED_print(char* str);
 
-void OLED_set_brightness(int lvl);
+void OLED_set_brightness(uint8_t lvl);
 
 
 #endif /* OLED_H_ */
